+++ Changelog.md
@@ -4,6 +4,7 @@
 
     - loggers : Create logging System
     - Lang : Create translate system
+    - Config : Create Config system
     - Entity : Can't have the same type of component two times
     
     - Crash Fix : Crash when show id of Entity Texts